refactor: util패키지 세부 분리
- parser 패키지 생성
- validator 패키지 생성
- generator 패키지 생성
