+++ src/main/java/lotto/model/domain/WinningLottoNumbers.java
@@ -1,0 +1,4 @@
+package lotto.model.domain;
+
+public class WinningLottoNumbers {
+}
