+++ src/main/java/lotto/util/generator/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package lotto.util;
+package lotto.util.generator;
 
 public class RandomNumberGenerator {
 }
