+++ src/main/java/lotto/util/validator/AmountValidator.java
@@ -1,4 +1,4 @@
-package lotto.util;
+package lotto.util.validator;
 
 import lotto.exception.InputErrorMessage;
