+++ src/test/java/lotto/util/parser/ParserTest.java
@@ -1,7 +1,6 @@
-package lotto.InputValidator;
+package lotto.util.parser;
 
 import lotto.exception.InputErrorMessage;
-import lotto.util.ParserUtil;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
@@ -12,7 +11,7 @@ public class ParserTest {
     @ParameterizedTest
     @ValueSource(strings = {"YOON", "*&!"})
     void 입력값이_숫자가_아니면_예외발생(String inValidAmount) {
-        assertThatThrownBy(() -> ParserUtil.parseInt(inValidAmount))
+        assertThatThrownBy(() -> ParserUtil.convertToNumber(inValidAmount))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage(InputErrorMessage.INVALID_NUMERIC_FORMAT.getMessage());
     }
@@ -20,7 +19,7 @@ public class ParserTest {
     @ParameterizedTest
     @ValueSource(strings = {""," "})
     void 입력값이_없다면_예외발생(String inValidAmount) {
-        assertThatThrownBy(() -> ParserUtil.parseInt(inValidAmount))
+        assertThatThrownBy(() -> ParserUtil.convertToNumber(inValidAmount))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage(InputErrorMessage.EMPTY_INPUT.getMessage());
     }
